requirements.txt
antlr4-python3-runtime
python-telegram-bot
matplotlib
networkx